Fixed GNU readline() input error due to bad handling of prompt variable
Before using GNU readline library
prompt = ">> "
During the usage of GNU readline library
prompt = prompt_font">> "input_font
where *_font contained ANSI CSI font format codes because of which the length
of prompt increased. This caused the readline library to concentrate on length
rather than what was printed on the screen which caused the input to repeat
from the beginning of the same line without completing the first line.
The fix,
rl_already_prompted=true;
was added in the init_readline() function and prompt was printed as was printed
before.